refactor: Board 패키지 분리 board -> main/heart/scrap (#32)
