--- src/main/java/paratrip/paratrip/board/controller/vo/response/BoardResponseVo.java
@@ -1,9 +1,0 @@
-package paratrip.paratrip.board.controller.vo.response;
-
-public class BoardResponseVo {
-	public record AddBoardResponse(
-		Long boardSeq
-	) {
-
-	}
-}
