+++ src/main/java/paratrip/paratrip/board/main/controller/vo/response/BoardResponseVo.java
@@ -1,0 +1,9 @@
+package paratrip.paratrip.board.main.controller.vo.response;
+
+public class BoardResponseVo {
+	public record AddBoardResponse(
+		Long boardSeq
+	) {
+
+	}
+}
