--- src/main/java/paratrip/paratrip/board/service/dto/request/BoardRequestDto.java
@@ -1,28 +1,0 @@
-package paratrip.paratrip.board.service.dto.request;
-
-import java.util.List;
-
-import org.springframework.web.multipart.MultipartFile;
-
-public class BoardRequestDto {
-	public record AddBoardRequestDto(
-		Long memberSeq,
-		String title,
-		String content,
-		String location,
-		List<MultipartFile> images
-	) {
-
-	}
-
-	public record ModifyBoardRequestDto(
-		Long memberSeq,
-		Long boardSeq,
-		String title,
-		String content,
-		String location,
-		List<MultipartFile> images
-	) {
-
-	}
-}
